requirements.txt
GitPython==3.1.24
python-gitlab==2.10.1
requests==2.14.2 # Same version as on the jenkins slaves
ruamel.yaml==0.17.10
